+++ rpctypes.h
@@ -24,14 +24,14 @@ enum rpctypes{
 
 struct __rpcbuff_el{
         struct __rpcbuff_el* childs;
-        size_t elen;
+        uint64_t elen;
         char* endpoint;
 };
 struct rpcbuff{
     struct __rpcbuff_el* start;
-    size_t* dimsizes;
-    size_t dimsizes_len;
-    size_t lastdim_len;
+    uint64_t* dimsizes;
+    uint64_t dimsizes_len;
+    uint64_t lastdim_len;
 };
 struct rpctype{
   char type;
@@ -39,11 +39,11 @@ struct rpctype{
   uint64_t datalen;
   char* data;
 };
-struct rpcbuff* rpcbuff_create(size_t* dimsizes,size_t dimsizes_len,size_t lastdim_len);  /*lastdim_len set default last dimension len(rpcbuff_getlast_from create this size lastdim if
+struct rpcbuff* rpcbuff_create(uint64_t* dimsizes,uint64_t dimsizes_len,uint64_t lastdim_len);  /*lastdim_len set default last dimension len(rpcbuff_getlast_from create this size lastdim if
                                                                                           this  endpoint wasnt writen by rpcbuff_pushto), but rpcbuff_pushto will fit bigger data*/
 void _rpcbuff_free(struct rpcbuff* rpcbuff);
-char* rpcbuff_getlast_from(struct  rpcbuff* rpcbuff, size_t* index, size_t index_len,size_t* outlen);
-int rpcbuff_pushto(struct rpcbuff* rpcbuff, size_t* index, size_t index_len, char* data, size_t data_len);
-char* rpcbuff_to_arr(struct rpcbuff* rpcbuff,size_t* buflen);
+char* rpcbuff_getlast_from(struct  rpcbuff* rpcbuff, uint64_t* index, uint64_t index_len,uint64_t* outlen);
+int rpcbuff_pushto(struct rpcbuff* rpcbuff, uint64_t* index, uint64_t index_len, char* data, uint64_t data_len);
+char* rpcbuff_to_arr(struct rpcbuff* rpcbuff,uint64_t* buflen);
 struct rpcbuff* buf_to_rpcbuff(char* buf);
-struct __rpcbuff_el* __rpcbuff_el_getlast_from(struct  rpcbuff* rpcbuff, size_t* index, size_t index_len);
+struct __rpcbuff_el* __rpcbuff_el_getlast_from(struct  rpcbuff* rpcbuff, uint64_t* index, uint64_t index_len);
